pyproject.toml
[project]
name = "lego-rag"
version = "1.0.0"
description = "A modern Retrieval-Augmented Generation (RAG) system for LEGO data"
authors = [
    {name = "LEGO RAG Team", email = "team@lego-rag.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.38.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "faiss-cpu>=1.7.4",
    "duckdb>=0.9.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0",
    "pydantic>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pytest",
    "pytest-cov",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["*.py"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.uv]
# Use faster resolver
resolution = "highest"
# Enable parallel downloads
index-strategy = "unsafe-best-match"
# Avoid CUDA dependencies
link-mode = "copy"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 